requirements.txt
Django==4.2
djangorestframework
requests
beautifulsoup4
